adk_calculator_sdk_from_json/requirements.txt
dbnl>=0.27.1
google-adk
google-generativeai
notebook